requirements.txt
psycopg2-binary
flask
flask-cors
Pillow
customtkinter
reportlab
google-generativeai
matplotlib
openpyxl
twilio
pandas
python-dotenv
tkcalendar
